gst-plugins-base: Autospec creation for update from version 1.20.3 to version 1.20.4
The latest bug-fix release in the 1.20 series is 1.20.4 and was released
on 12 October 2022.

See https://gstreamer.freedesktop.org/releases/1.20/ for the latest
version of this document.

Last updated: Wednesday 12 October 2022, 15:00 UTC (log)

Introduction

The GStreamer team is proud to announce a new major feature release in
the stable 1.x API series of your favourite cross-platform multimedia
framework!

As always, this release is again packed with many new features, bug

(NEWS truncated at 15 lines)